update(*): module 5-task2
